ml/requirements.txt
boto3==1.36.16
keras==3.3.0
paho-mqtt==2.1.0
pillow==10.4.0
python-dotenv==1.0.1
scikit-learn==1.5.2
tensorflow==2.17.0
wandb==0.18.1